% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aec2013_2cp_electorate}
\alias{aec2013_2cp_electorate}
\title{2013 General election data for two candidate preferred votes for candidates for the House of Representative for each electorate}
\format{A data frame with the following variables:
\itemize{
    \item ID: Electoral division ID    
    \item Electorate:  Electoral division name   
    \item CandidateID: Candidate ID       
    \item Surname: Candidate surname          
    \item GivenNm: Candidate given name            
    \item BallotPosition: Candidate's position on the ballot    
    \item Elected: Whether the candidate was elected (Y/N)           
    \item HistoricElected:   
    \item PartyAb: Abbreviation for political party name           
    \item PartyNm: Political party name           
    \item OrdinaryVotes: Number of ordinates votes cast for the candidate     
    \item AbsentVotes: Number of absentee votes cast for the candidate
    \item ProvisionalVotes: Number of provisional votes cast for the candidate
    \item PrePollVotes: Number of pre-poll votes cast for the candidate
    \item PostalVotes: Number of postal votes cast for the candidate
    \item TotalVotes: Total votes cast for the candidate
    \item Swing: Percentage change since the last election
    }}
\usage{
aec2013_2cp_electorate
}
\description{
A dataset containing two candidate preferred vote counts, and other results for the House of Representatives from the 2013 Australian federal election. Includes the count of votes for
the leading two candidates in the electorate after distribution of preferences.
The data were obtained from the Australian Electoral Commission, and downloaded 
from \url{http://results.aec.gov.au/17496/Website/HouseDownloadsMenu-17496-csv.htm} and 
\url{http://www.aec.gov.au/elections/federal_elections/2013/downloads.htm}.
}
\keyword{datasets}

