% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aec2013_2pp_electorate}
\alias{aec2013_2pp_electorate}
\title{2013 General election data for two party preferred votes for candidates for the House of Representative for each electorate}
\format{A data frame with the following variables:
\itemize{
    \item ID: Electoral division ID    
    \item Electorate:  Electoral division name   
    \item CandidateID: Candidate ID       
    \item Surname: Candidate surname          
    \item GivenNm: Candidate given name            
    \item BallotPosition: Candidate's position on the ballot    
    \item Elected: Whether the candidate was elected (Y/N)           
    \item HistoricElected:   
    \item PartyAb: Abbreviation for political party name           
    \item PartyNm: Political party name           
    \item OrdinaryVotes: Number of ordinates votes cast at the polling place for the candidate     
    \item State: Abbreviation for state name        
    \item Total_Australian_Labor_Party_Votes_per_electorate:     
    \item Average_Australian_Labor_Party_Percentage_in_electorate:    
    \item Total_Liberal_National_Coalition_Votes_per_electorate:      
    \item Average_Liberal_National_Coalition_Percentage_in_electorate:
    \item Total_2pp_votes_per_electorate:     
    }}
\usage{
aec2013_2pp_electorate
}
\description{
A dataset containing two party preferred vote counts, winning candidate names, and other results for the House of Representatives from the 2013 Australian federal election. Includes the count of votes for
the Australian Labor Party and the count of votes for the Liberal-National Coalition for each polling place.
The data were obtained from the Australian Electoral Commission, and downloaded 
from \url{http://results.aec.gov.au/17496/Website/HouseDownloadsMenu-17496-csv.htm} and 
\url{http://www.aec.gov.au/elections/federal_elections/2013/downloads.htm}.
}
\keyword{datasets}

