% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aec2013_electorates}
\alias{aec2013_electorates}
\title{Electorate-level 2013 summary of the general election data for the House of Representatives}
\format{A data frame with the following variables:
\itemize{
    \item ID: Electoral division ID    
    \item Electorate:  Electoral division name   
    \item CandidateID: Candidate ID       
    \item Surname: Candidate surname          
    \item GivenNm: Candidate given name            
    \item BallotPosition: Candidate's position on the ballot    
    \item Elected: Whether the candidate was elected (Y/N)           
    \item HistoricElected:   
    \item PartyAb: Abbreviation for political party name           
    \item PartyNm: Political party name           
    \item OrdinaryVotes: Number of ordinates votes cast at the polling place for the candidate     
    \item State: Abbreviation for state name             
    }}
\usage{
aec2013_electorates
}
\description{
A dataset containing vote counts, candidate names, and other national,
 state, and divisional results for the
House of Representatives from the 2013 Australian federal election. 
The data were obtained from the Australian Electoral Commission, and downloaded 
from \url{http://results.aec.gov.au/17496/Website/HouseDownloadsMenu-17496-csv.htm} and 
\url{http://www.aec.gov.au/elections/federal_elections/2013/downloads.htm}.
}
\keyword{datasets}

