% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartogram.R
\name{aec_carto_f}
\alias{aec_carto_f}
\title{aec_carto_f - run dorling ondata centers}
\usage{
aec_carto_f(aec_data_sub, polygon.vertex = 6, name.text = TRUE,
  dist.ratio = dist.ratio, iteration = 100, xlab = "", ylab = "", ...)
}
\arguments{
\item{aec_data_sub}{subset of data with centroids of electoral divisions}

\item{...}{arguments to dorling function}
}
\description{
The dorling algorithm creates a non-contiguous cartogram by
shifting circles to alleviate overlap, while roughly maintaining
geographic proximity.
}
\examples{
data(nat_map)
data(nat_data)
aec_data_syd <- aec_extract(nat_data, expand=list(c(2,3)))
aec_data_dor <- aec_carto(aec_data_syd, expand=list(c(3,4.5)))
}

