% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartogram.R
\name{complete_color}
\alias{complete_color}
\title{Auto complete (or cut) a vector to a fixed length}
\usage{
complete_color(cl, targetlen)
}
\arguments{
\item{cl}{a vector}

\item{targetlen}{the target length}
}
\value{
a vector of completed cl with length n
}
\description{
From https://github.com/chxy/cartogram/blob/master/R/dorling.R
Not exported here, but needed for aec_carto_f
}
\examples{
\dontrun{
complete_color('red',5)
complete_color(c('red','blue'),5)
complete_color(c('red','blue','green','yellow','pink','grey'),5)
}
}

