% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles.R
\name{download_ShapeFile}
\alias{download_ShapeFile}
\title{Download a shapefile from the aec website}
\usage{

  download_ShapeFile(url = "http://www.aec.gov.au/Electorates/gis/gis_datadownload.htm",
  exdir = "temp", stem = "http://www.aec.gov.au/Electorates/gis/",
  keep = FALSE)
}
\arguments{
\item{url}{url of aec website}

\item{exdir}{relative path of folder where shapefile should be downloaded to}

\item{stem}{base directory url of aec website}
}
\value{
object of class SpatialPolygonsDataFrame
}
\description{
Download a shapefile from the aec website
}
\examples{
\dontrun{
x <- download_ShapeFile(exdir = "Shapefiles")
# user input 21
sFsmall <- rmapshaper::ms_simplify(x, keep=0.05) # use instead of thinnedSpatialPoly
plot(sFsmall)
}
}

