\name{eechidna-package}
\alias{eechidna-package}
\alias{eechidna}
\docType{package}
\title{Exploring Election and Census Highly Informative Data Nationally for Australia}

\description{Data from 2013 Australian Federal Election and 2011 Australian
    Census for each House of Representatives electorate, along with some tools
    for visualizing and analysing the data.
}

\author{
    Di Cook, 
    Heike Hofmann, 
    Rob Hyndman, 
    Thomas Lumley,
    Ben Marwick, 
    Carson Sievert,
    Nicholas Tierney,
    Nathaniel Tomasetti, 
    Fang Zhou.
}


\keyword{package}

