% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{extract_centroids}
\alias{extract_centroids}
\title{Extract centroids from the polygons within a shapefile.}
\usage{
extract_centroids(shapefile)
}
\arguments{
\item{shapefile}{sf object containing polygon geometry}
}
\value{
list containing centroids as sf points object and a dataframe with
basic data on each polygon (e.g. name)
}
\description{
Extract centroids from the polygons within a shapefile.
}
\examples{
\dontrun{
sF_16 <- sF_download(year = 2016)
electorate_centroids_2016 <- extract_centroids(sF_16)
}
}
