% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fp01}
\alias{fp01}
\title{2001 General election data for first preference votes for candidates for the House of Representative for each electorate}
\format{A data frame with the following variables:
\itemize{
    \item StateAb: Abbreviation for state name  
    \item DivisionNm:  Electoral division name   
    \item Surname: Candidate surname          
    \item GivenNm: Candidate given name            
    \item PartyAb: Abbreviation for political party name           
    \item PartyNm: Political party name      
    \item Elected: Whether the candidate was elected (Y/N)            
    \item Percent: Percentage of ordinary votes for the candidate     
    }}
\usage{
fp01
}
\description{
A dataset containing first preference vote counts, candidate names, and other results for the
House of Representatives from the 2001 Australian federal election. 
The data were obtained from the Australian Electoral Commission.
}
\keyword{datasets}
