% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fp_pp01}
\alias{fp_pp01}
\title{2001 General election data for first preference votes for candidates for the House of Representatives for each polling place}
\format{A data frame with the following variables:
\itemize{
    \item StateAb: Abbreviation for state name  
    \item DivisionNm:  Electoral division name   
    \item PollingPlace: Polling place name     
    \item Surname: Candidate surname          
    \item GivenNm: Candidate given name            
    \item BallotPosition: Candidate's position on the ballot         
    \item PartyAb: Abbreviation for political party name    
    \item Elected: Whether the candidate was elected (Y/N)  
    \item OrdinaryVotes: Number of ordinary votes cast at the polling place for the candidate  
    \item Latitude: Coordinates        
    \item Longitude: Coordinates
    }}
\usage{
fp_pp01
}
\description{
A dataset containing first preference vote counts, candidate names, polling place locations,
and other results for the
House of Representatives from the 2001 Australian federal election. 
The data were obtained from the Australian Electoral Commission.
}
\keyword{datasets}
