% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fp_pp07}
\alias{fp_pp07}
\title{2007 General election data for first preference votes for candidates for the House of Representatives for each polling place}
\format{A data frame with the following variables:
\itemize{
    \item StateAb: Abbreviation for state name    
    \item DivisionID: Electoral division ID    
    \item DivisionNm:  Electoral division name   
    \item PollingPlaceID: Polling place ID  
    \item PollingPlace: Polling place name     
    \item CandidateID: Candidate ID       
    \item Surname: Candidate surname          
    \item GivenNm: Candidate given name            
    \item BallotPosition: Candidate's position on the ballot    
    \item Elected: Whether the candidate was elected (Y/N)           
    \item HistoricElected: Whether the candidate is the incumbent member  
    \item PartyAb: Abbreviation for political party name           
    \item PartyNm: Political party name           
    \item OrdinaryVotes: Number of ordinary votes cast at the polling place for the candidate     
    \item Swing: Percentage point change in ordinary votes for the party from the previous election     
    \item PremisesPostCode: Post code of polling booth  
    \item Latitude: Coordinates        
    \item Longitude: Coordinates
    }}
\usage{
fp_pp07
}
\description{
A dataset containing first preference vote counts, candidate names, polling place locations,
and other results for the
House of Representatives from the 2007 Australian federal election. 
The data were obtained from the Australian Electoral Commission, and downloaded 
from \url{http://results.aec.gov.au/13745/Website/HouseDownloadsMenu-13745-csv.htm} and 
\url{http://www.aec.gov.au/elections/federal_elections/2007/downloads.htm}.
}
\keyword{datasets}
