% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hexDat}
\alias{hexDat}
\title{Electorate hexagon data in a tidy form}
\format{A data frame}
\usage{
hexDat
}
\description{
Electorate hexagon data in a tidy form
}
\examples{
data(hexDat)
library(plotly)
p <- ggplot(hexDat, aes(xcent, ycent, text = Electorate)) + 
  geom_hex(stat = "identity") + 
  lims(x=c(-80, 8), y=c(-40, 50))
ggplotly(p, tooltip = "text")
}
\references{
Thomas Lumley
}
\seealso{
Thomas Lumley
}
\keyword{datasets}

