% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{launchApp}
\alias{launchApp}
\title{Shiny app for exploring census and electorate data}
\usage{
launchApp(ageVars = c("Age00_04", "Age05_14", "Age15_19", "Age20_24",
  "Age25_34", "Age35_44", "Age45_54", "Age55_64", "Age65_74", "Age75_84",
  "Age85plus"))
}
\description{
Shiny app for exploring census and electorate data
}
\examples{
\dontrun{
launchApp()
}
}

