% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles.R
\name{loadShapeFile}
\alias{loadShapeFile}
\title{Load shapefile of Australia into R}
\usage{
loadShapeFile(path_to_shapeFile, tolerance = 0.005)
}
\arguments{
\item{path_to_shapeFile}{path to object in local machine}

\item{tolerance}{numerical tolerance value to be used by the Douglas-Peuker algorithm}
}
\value{
object of class SpatialPolygonsDataFrame
}
\description{
Load shapefile into R as a SpatialDataFrame, extract polygon information, thin
polygon, fix any problematic polygons, and format variable names.
"nat_map" and "nat_data" objects for every Australian federal election between 
2001-2016 can be readily loaded from the package for analysis.
}
\details{
The function will take several minutes to complete.
}
\examples{
\dontrun{
fl <- "data-raw/national-midmif-09052016/COM_ELB.TAB"
sF_2016 <- loadShapeFile("local/path/to/shapefile.shp")
}
}
