% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nat_data01}
\alias{nat_data01}
\title{Data of the Australian Electorates from 2001}
\format{A data frame with 150 rows with the following variables:
\itemize{
    \item id: numeric identifier that links the electorate with the corresponding polygon in `nat_map`.
    \item elect_div: Electorate division name   
    \item state: abbreviation of the state name
    \item long_c: longitude coordinate of electorate (polygon) centroid
    \item lat_c: latitude coordinate of electorate (polygon) centroid
    \item x: latitude coordinate for plotting a cartogram
    \item y: longitude coordinate for plotting a cartogram
    \item radius: variable used in the construction of cartogram points
}}
\usage{
nat_data01
}
\description{
Data of the Australian Electorates from 2001
}
\keyword{datasets}
