% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nat_data04}
\alias{nat_data04}
\title{Data of the Australian Electorates from 2004}
\format{A data frame with 150 rows with the following variables:
\itemize{
    \item UniqueID: numeric identifier that links the electorate with the corresponding polygon 
    in `nat_map` and other data sets.
    \item elect_div: Electorate division name   
    \item state: abbreviation of the state name
    \item long_c: longitude coordinate of electorate (polygon) centroid
    \item lat_c: latitude coordinate of electorate (polygon) centroid
    \item x: latitude coordinate for plotting a cartogram
    \item y: longitude coordinate for plotting a cartogram
    \item radius: variable used in the construction of cartogram points
}}
\usage{
nat_data04
}
\description{
A dataset containing some demographic information for each of the 150 Australian electorates.
The data were obtained from the Australian Bureau of Statistics, and downloaded 
from \url{http://www.abs.gov.au/AUSSTATS/abs@.nsf/DetailsPage/2923.0.30.0012006?OpenDocument}.
The data is published
}
\keyword{datasets}
