% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nat_data07}
\alias{nat_data07}
\title{Data and centroids corresponding to the Australian Electorates from 2007}
\format{
A data frame with 150 rows with the following variables:
\itemize{
    \item id: Numeric identifier for the polygon
    \item elect_div: Electorate division name
    \item state: abbreviation of the state name
    \item long_c: longitude coordinate of electorate (polygon) centroid
    \item lat_c: latitude coordinate of electorate (polygon) centroid
    \item x: latitude coordinate for plotting a cartogram
    \item y: longitude coordinate for plotting a cartogram
    \item radius: variable used in the construction of cartogram points
}
}
\usage{
nat_data07
}
\description{
A dataset containing some demographic information for each of the 150 Australian electorates.
The data were obtained from the Australian Bureau of Statistics, and downloaded
from \url{https://www.abs.gov.au/AUSSTATS/abs@.nsf/DetailsPage/2923.0.30.0012006?OpenDocument}.
Must be loaded using the `nat_data_download` function.
}
\keyword{internal}
