% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nat_data19}
\alias{nat_data19}
\title{Data and centroids corresponding to the Australian Electorates from 2019}
\format{
A data frame with 151 rows with the following variables:
\itemize{
    \item id: Numeric identifier for the polygon
    \item elect_div: Electorate division name   
    \item state: abbreviation of the state name
    \item numccds: AEC variable that might be filled with meaning or a description down the road
    \item area_sqkm: combined square kilometers of each electorate
    \item long_c: longitude coordinate of electorate (polygon) centroid
    \item lat_c: latitude coordinate of electorate (polygon) centroid
    \item x: latitude coordinate for plotting a cartogram
    \item y: longitude coordinate for plotting a cartogram
    \item radius: variable used in the construction of cartogram points
}
}
\usage{
nat_data19
}
\description{
A dataset containing some demographic information for each of the 151 Australian electorates.
The data were obtained from the Australian Electoral Commission, and downloaded 
from \url{https://www.aec.gov.au/Electorates/gis/gis_datadownload.htm}.
Must be loaded using the `nat_data_download` function.
}
\keyword{internal}
