% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nat_data_2016_cart}
\alias{nat_data_2016_cart}
\title{This is the same as nat_data_2016, except that it has 
spatial centroids of electorates spread out using the 
Dorling cartogram.}
\format{An object of class \code{data.frame} with 153 rows and 10 columns.}
\usage{
nat_data_2016_cart
}
\description{
This is the same as nat_data_2016, except that it has 
spatial centroids of electorates spread out using the 
Dorling cartogram.
}
\keyword{datasets}
