% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nat_data_cart}
\alias{nat_data_cart}
\title{This is the same as nat_data, but with a few more cols. 
We can probably replace nat_dat with this one}
\format{An object of class \code{data.frame} with 150 rows and 10 columns.}
\usage{
nat_data_cart
}
\description{
This is the same as nat_data, but with a few more cols. 
We can probably replace nat_dat with this one
}
\keyword{datasets}

