% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nat_map01}
\alias{nat_map01}
\title{Map of Australian Electorates from 2001}
\format{An object of class \code{data.frame} with 27025 rows and 9 columns.}
\usage{
nat_map01
}
\description{
Map of Australian Electorates from 2001
}
\keyword{datasets}
