% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nat_map07}
\alias{nat_map07}
\title{Map of Australian Electorates from 2007}
\format{
A data frame with the following variables:
\itemize{
    \item id: Numeric identifier for the polygon
    \item long: longitude coordinate of point in polygon
    \item lat: latitude coordinate of point in polygon
    \item order: order for polygon points
    \item hole: whether polygon has a hole
    \item piece: piece for polygon
    \item group: group for polygon
    \item elect_div: Electoral division name
    \item state: Abbreviation for state name
}
}
\usage{
nat_map07
}
\description{
A dataset containing the map of the all 150 Australian electorates using the 2007 boundaries of the 
electorates. The data were obtained from the Australian Electoral Commission, and downloaded 
from \url{http://www.abs.gov.au/AUSSTATS/abs@.nsf/DetailsPage/2923.0.30.0012006?OpenDocument}. 
Must be loaded using the `nat_map_download` function.
}
\keyword{internal}
