% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nat_map16}
\alias{nat_map16}
\title{--------------------------------------------------------------------------------------------------
Map of Australian Electorates from 2016}
\format{An object of class \code{data.frame} with 48897 rows and 9 columns.}
\usage{
nat_map16
}
\description{
A dataset containing the map of the all 150 Australian electorates using the 2016 boundaries of the 
electorates (and downsampled to a 5\% file to allow fast plotting).
The data were obtained from the Australian Electoral Commission, and downloaded 
from \url{http://www.aec.gov.au/Electorates/gis/gis_datadownload.htm}.
}
\examples{

data(nat_map16)
# choropleth map with Census data
nat_map16$region <- nat_map16$elect_div
data(abs2016)
abs2016$region <- abs2016$Electorate
library(ggplot2)
library(ggthemes)
both <- intersect(unique(abs2016$region), unique(nat_map16$region))
ggplot(aes(map_id=region), data=subset(abs2016, region \%in\% both)) +
  geom_map(aes(fill=MedianIncome), map=subset(nat_map16, region \%in\% both)) +
  expand_limits(x=nat_map16$long, y=nat_map16$lat) + 
  theme_map()

}
\keyword{datasets}
