% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sF_07}
\alias{sF_07}
\title{Shapefile containing Australian Electorates from 2007}
\format{An object of class \code{SpatialPolygonsDataFrame} with 150 rows and 5 columns.}
\usage{
sF_07
}
\description{
A large SpatialPolygonsDataFrame containing the polygons and associated data for each of the 150 Australian electorates in the 2007 federal election.
The data were obtained from the Australian Electoral Commission, and downloaded 
from \url{http://www.abs.gov.au/AUSSTATS/abs@.nsf/DetailsPage/2923.0.30.0012006?OpenDocument}.
}
\examples{

}
\keyword{datasets}
