% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sF_11}
\alias{sF_11}
\title{Shapefile containing Australian Electorates in place during 2011}
\format{An object of class \code{SpatialPolygonsDataFrame} with 150 rows and 5 columns.}
\usage{
sF_11
}
\description{
A large SpatialPolygonsDataFrame containing the polygons and associated data for each of the 150 Australian electorates at the time of the 2011 Census.
The data were obtained from the Australian Bureau of Statistics, and downloaded 
from \url{https://datapacks.censusdata.abs.gov.au/datapacks/}.
}
\keyword{datasets}
