% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sF_16}
\alias{sF_16}
\title{Shapefile containing Australian Electorates from 2016}
\format{An object of class \code{SpatialPolygonsDataFrame} with 150 rows and 11 columns.}
\usage{
sF_16
}
\description{
A large SpatialPolygonsDataFrame containing the polygons and associated data for each of the 150 Australian electorates in the 2016 federal election.
The data were obtained from the Australian Electoral Commission, and downloaded 
from \url{http://www.aec.gov.au/Electorates/gis/gis_datadownload.htm}.
}
\keyword{datasets}
