% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcp16}
\alias{tcp16}
\title{2016 General election data for two candidate preferred votes for candidates for the House of Representative for each electorate}
\format{A data frame with the following variables:
\itemize{
    \item StateAb: Abbreviation for state name    
    \item DivisionID: Electoral division ID    
    \item DivisionNm:  Electoral division name             
    \item BallotPosition: Candidate's position on the ballot      
    \item CandidateID: Candidate ID       
    \item Surname: Candidate surname          
    \item GivenNm: Candidate given name 
    \item PartyAb: Abbreviation for political party name           
    \item PartyNm: Political party name        
    \item Elected: Whether the candidate was elected (Y/N)           
    \item HistoricElected: Whether the candidate is the incumbent member      
    \item OrdinaryVotes: Number of ordinary votes cast for the candidate 
    \item Percent: Percentage of ordinary votes cast for the candidate    
    }}
\usage{
tcp16
}
\description{
A dataset containing two candidate preferred vote counts, and other results for the House of Representatives from the 2016 Australian federal election. Includes the count of votes for
the leading two candidates in the electorate after distribution of preferences.
The data were obtained from the Australian Electoral Commission, and downloaded 
from \url{http://results.aec.gov.au/20499/Website/HouseDownloadsMenu-20499-csv.htm} and 
\url{http://www.aec.gov.au/elections/federal_elections/2016/downloads.htm}.
}
\keyword{datasets}
