% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tpp}
\alias{tpp}
\alias{tpp01}
\alias{tpp04}
\alias{tpp10}
\alias{tpp07}
\alias{tpp13}
\alias{tpp16}
\alias{tpp19}
\alias{tpp21}
\title{Two party preferred votes for candidates in each electorate where Labor and
Liberal parties were the two most popular parties.}
\format{
A data frame with the following variables: 
\describe{ 
  \item{UniqueID}{Numeric identifier that links the electoral division with Census and other election datasets.}
  \item{DivisionNm}{Electoral division name}
  \item{StateAb}{Abbreviation for state name}
  \item{LNP_Vote}{Count of two party preferred vote in favour of the Liberal National coalition}
  \item{LNP_Percent}{Percentage of two party preferred vote in favour of the Liberal National coalition}
  \item{ALP_Votes}{Count of two party preferred vote in favour of the Labor party}
  \item{ALP_Percent}{Percentage of two party preferred vote in favour of the Labor party}
  \item{TotalVotes}{Total number of votes cast}
  \item{Swing}{Percentage point change in two party preferred vote from the previous election}
 }

{}

{}

{}

{}

{}

{}

{}
}
\source{
\url{https://www.aec.gov.au/About_AEC/Publications/statistics/files/aec-2001-election-statistics.zip}

\url{https://results.aec.gov.au/12246/results/HouseDownloadsMenu-12246-csv.htm}

\url{https://www.aec.gov.au/elections/federal_elections/2004/downloads.htm}

\url{https://results.aec.gov.au/15508/Website/HouseDownloadsMenu-15508-csv.htm}

\url{https://www.aec.gov.au/elections/federal_elections/2010/downloads.htm}

\url{https://results.aec.gov.au/13745/Website/HouseDownloadsMenu-13745-csv.htm}

\url{https://www.aec.gov.au/elections/federal_elections/2007/downloads.htm}

\url{https://results.aec.gov.au/17496/Website/HouseDownloadsMenu-17496-csv.htm}

\url{https://www.aec.gov.au/elections/federal_elections/2013/downloads.htm}

\url{https://results.aec.gov.au/20499/Website/HouseDownloadsMenu-20499-csv.htm}

\url{https://www.aec.gov.au/elections/federal_elections/2016/downloads.htm}

\url{https://results.aec.gov.au/24310/Website/HouseDownloadsMenu-24310-Csv.htm}

\url{https://results.aec.gov.au/27966/Website/HouseDownloadsMenu-27966-Csv.htm}
}
\usage{
tpp01

tpp04

tpp10

tpp07

tpp13

tpp16

tpp19

tpp21
}
\description{
A dataset containing two party preferred vote counts, winning candidate
names, and other results for the House of Representatives from the 2001 to
2019 Australian federal elections. Includes the count of votes for the
Australian Labor Party and the count of votes for the Liberal-National
Coalition for each electorate.
}
\keyword{datasets}
