% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tpp_pp01}
\alias{tpp_pp01}
\title{2001 General election data for two party preferred votes for the House of Representatives for each polling place}
\format{A data frame with the following variables:
\itemize{
    \item StateAb: Abbreviation for state name 
    \item DivisionNm:  Electoral division name 
    \item PollingPlace: Polling place name  
    \item LNP_Votes: Count of two party preferred vote in favour of the Liberal National coalition
    \item LNP_Percent: Percentage of two party preferred vote in favour of the Liberal National coalition   
    \item ALP_Votes: Count of two party preferred vote in favour of the Labor party
    \item ALP_Percent: Percentage of two party preferred vote in favour of the Labor party
    \item TotalVotes: Total number of votes cast   
    \item Swing: Percentage point change in ALP percentage two party preferred vote from the previous election   
    \item Latitude: Coordinates        
    \item Longitude: Coordinates    
    }}
\usage{
tpp_pp01
}
\description{
A dataset containing two party preferred vote counts, winning candidate names, polling place locations,
and other results for the House of Representatives from the 2001 Australian federal election. Includes the count of votes for
the Australian Labor Party and the count of votes for the Liberal-National Coalition for each polling place.
The data were obtained from the Australian Electoral Commission.
}
\keyword{datasets}
