% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aec2013}
\alias{aec2013}
\title{2013 General election data for the House of Representatives}
\format{An object of class \code{tbl_df} (inherits from \code{data.frame}) with 82398 rows and 28 columns.}
\usage{
aec2013
}
\description{
A dataset containing vote counts, candidate names, polling place locations,
and other national, state, divisional and polling place results for the
House of Representatives from the 2013 Australian federal election. 
The data were obtained from the Australian Electoral Commission, and downloaded 
from \url{http://results.aec.gov.au/17496/Website/HouseDownloadsMenu-17496-csv.htm} and 
\url{http://www.aec.gov.au/elections/federal_elections/2013/downloads.htm}.
}
\keyword{datasets}

