% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartogram.R
\name{aec_carto_f}
\alias{aec_carto_f}
\title{aec_carto_f - run dorling ondata centers}
\usage{
aec_carto_f(aec_data_sub, polygon.vertex = 6, name.text = TRUE,
  dist.ratio = dist.ratio, iteration = 100, xlab = "", ylab = "", ...)
}
\arguments{
\item{aec_data_sub}{subset of data with centroids of electoral divisions}

\item{polygon.vertex}{x}

\item{name.text}{x}

\item{dist.ratio}{x}

\item{iteration}{x}

\item{xlab}{x}

\item{ylab}{x}

\item{...}{arguments to dorling function}
}
\description{
The dorling algorithm creates a non-contiguous cartogram by
shifting circles to alleviate overlap, while roughly maintaining
geographic proximity.
}

