\name{echidnaR-package}
\alias{echidnaR-package}
\alias{echidnaR}
\docType{package}
\title{Electoral Commission Handling Intelligent Data Nationally for Australia using R}

\description{Data from 2013 Australian Federal Election and 2011 Australian
    Census for each House of Representatives electorate, along with some tools
    for visualizing and analysing the data.
}

\author{
    Di Cook, 
    Heike Hoffman, 
    Rob Hyndman, 
    Ben Marwick, 
    Carson Sievert,
    Nicholas Tierney, 
    Fang Zhou.
}


\keyword{package}

