% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fp01}
\alias{fp01}
\title{2001 Australian Federal election data: First preference votes for candidates (House of Representatives) in each electorate.}
\format{A data frame with the following variables:
\itemize{
    \item UniqueID: Numeric identifier that links the electoral division with Census  
    and other election datasets.
    \item StateAb: Abbreviation for state name  
    \item DivisionNm:  Electoral division name   
    \item Surname: Candidate surname          
    \item GivenNm: Candidate given name            
    \item PartyAb: Abbreviation for political party name           
    \item PartyNm: Political party name      
    \item Elected: Whether the candidate was elected (Y/N)            
    \item Percent: Percentage of ordinary votes for the candidate     
    }}
\usage{
fp01
}
\description{
A dataset containing first preference vote counts, candidate names, and other results for the
House of Representatives from the 2001 Australian federal election. 
The data were obtained from the Australian Electoral Commission, and downloaded 
from \url{https://www.aec.gov.au/About_AEC/Publications/statistics/files/aec-2001-election-statistics.zip}.
}
\keyword{datasets}
