% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fp_pp01}
\alias{fp_pp01}
\title{--------------------------------------------------------------------------------------------------
2001 General election data for first preference votes for candidates for the House of Representatives for each polling place}
\format{A data frame with the following variables:
\itemize{
    \item StateAb: Abbreviation for state name     
    \item DivisionNm:  Electoral division name   
    \item PollingPlace: Polling place name  
    \item Latitude: Coordinates        
    \item Longitude: Coordinates
    \item TotalVotes: Total number of votes submitted at polling booth
    \item ALP: Percentage of votes for the Labor party
    \item GRN: Percentage of votes for the Greens
    \item ON: Percentage of votes for One Nation
    \item IND: Percentage of votes for independent candidates
    \item LNP: Percentage of votes for the Liberal National Coalition  
    \item Other: Percentage of votes for other parties
    }}
\usage{
fp_pp01
}
\description{
A dataset containing first preference vote counts, candidate names, polling place locations,
and other results for the
House of Representatives from the 2001 Australian federal election. 
The data were obtained from the Australian Electoral Commission.
}
\keyword{datasets}
