% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{launchApp}
\alias{launchApp}
\title{Shiny app for exploring census and electorate data}
\usage{
launchApp()
}
\description{
Shiny app for exploring census and electorate data
}
\examples{
\dontrun{
launchApp()
}
}

