% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nat_map01}
\alias{nat_map01}
\title{Map of Australian Electorates from 2001}
\format{An object of class \code{data.frame} with 38633 rows and 9 columns.}
\usage{
nat_map01
}
\description{
A dataset containing the map of the all 150 Australian electorates using the 2001 boundaries of the 
electorates (and downsampled to a 5\% file to allow fast plotting).
The data were obtained from the Australian Government, and downloaded 
from \url{https://data.gov.au/dataset/ds-dga-0b939a62-e53e-4616-add5-77f909b58ddd/details?q=asgc\%202001}.
}
\keyword{datasets}
