% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nat_map07}
\alias{nat_map07}
\title{Map of Australian Electorates from 2010}
\format{An object of class \code{data.frame} with 31955 rows and 9 columns.}
\usage{
nat_map07

nat_map07
}
\description{
A dataset containing the map of the all 150 Australian electorates using the 2010 boundaries of the 
electorates (and downsampled to a 5\% file to allow fast plotting).
The data were obtained from the Australian Electoral Commission, and downloaded 
from \url{http://www.aec.gov.au/Electorates/gis/gis_datadownload.htm}.

A dataset containing the map of the all 150 Australian electorates using the 2007 boundaries of the 
electorates (and downsampled to a 5\% file to allow fast plotting).
The data were obtained from the Australian Electoral Commission, and downloaded 
from \url{http://www.abs.gov.au/AUSSTATS/abs@.nsf/DetailsPage/2923.0.30.0012006?OpenDocument}.
}
\keyword{datasets}
